Refactored Scope Secret operations to ReconConfigPrompts
