+++ tests/unit/helpers/test_recon_config_utils.py
@@ -1,19 +1,21 @@
-# pylint: disable=wrong-import-order,ungrouped-imports,useless-suppression
 from unittest.mock import patch
 
 import pytest
 from databricks.labs.blueprint.tui import MockPrompts
+from databricks.sdk.errors.platform import ResourceDoesNotExist
+from databricks.sdk.service.workspace import SecretScope
 
 from databricks.labs.remorph.helpers.recon_config_utils import ReconConfigPrompts
 
+SOURCE_DICT = {"databricks": "0", "netezza": "1", "oracle": "2", "snowflake": "3"}
+SCOPE_NAME = "dummy_scope"
+
 
 def test_configure_secrets_snowflake_overwrite(mock_workspace_client):
-    source_dict = {"databricks": "0", "netezza": "1", "oracle": "2", "snowflake": "3"}
-    scope_name = "dummy_scope"
     prompts = MockPrompts(
         {
-            r"Select the source": source_dict["snowflake"],
-            r"Enter Secret Scope name": scope_name,
+            r"Select the source": SOURCE_DICT["snowflake"],
+            r"Enter Secret Scope name": SCOPE_NAME,
             r"Enter Snowflake URL": "dummy",
             r"Enter Account Name": "dummy",
             r"Enter User": "dummy",
@@ -25,25 +27,20 @@ def test_configure_secrets_snowflake_overwrite(mock_workspace_client):
             r"Do you want to overwrite.*": "yes",
         }
     )
+    mock_workspace_client.secrets.list_scopes.side_effect = [[SecretScope(name=SCOPE_NAME)]]
+    recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
+    recon_conf.prompt_source()
 
-    with patch(
-        "databricks.labs.remorph.helpers.db_workspace_utils.DatabricksSecretsClient._scope_exists",
-        return_value=True,
-    ):
-        recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
-        recon_conf.prompt_source()
-
-        recon_conf.prompt_and_save_connection_details()
+    recon_conf.prompt_and_save_connection_details()
 
 
 def test_configure_secrets_oracle_insert(mock_workspace_client):
-    source_dict = {"databricks": "0", "netezza": "1", "oracle": "2", "snowflake": "3"}
-    scope_name = "dummy_scope"
     # mock prompts for Oracle
     prompts = MockPrompts(
         {
-            r"Select the source": source_dict["oracle"],
-            r"Enter Secret Scope name": scope_name,
+            r"Select the source": SOURCE_DICT["oracle"],
+            r"Enter Secret Scope name": SCOPE_NAME,
+            r"Do you want to create a new one?": "yes",
             r"Enter User": "dummy",
             r"Enter Password": "dummy",
             r"Enter host": "dummy",
@@ -52,34 +49,28 @@ def test_configure_secrets_oracle_insert(mock_workspace_client):
         }
     )
 
-    with patch(
-        "databricks.labs.remorph.helpers.db_workspace_utils.DatabricksSecretsClient._scope_exists",
-        return_value=True,
-    ):
-        # mock secret_key_exists to return True
-        with patch(
-            "databricks.labs.remorph.helpers.db_workspace_utils.DatabricksSecretsClient.secret_key_exists",
-            return_value=False,
-        ):
-            recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
-            recon_conf.prompt_source()
+    mock_workspace_client.secrets.list_scopes.side_effect = [[SecretScope(name="scope_name")]]
 
-            recon_conf.prompt_and_save_connection_details()
+    with patch(
+        "databricks.labs.remorph.helpers.recon_config_utils.ReconConfigPrompts._secret_key_exists",
+        return_value=False,
+    ):
+        recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
+        recon_conf.prompt_source()
+
+        recon_conf.prompt_and_save_connection_details()
 
 
 def test_configure_secrets_invalid_source(mock_workspace_client):
-    source_dict = {"databricks": "0", "netezza": "1", "oracle": "2", "snowflake": "3"}
-    scope_name = "dummy_scope"
-    # mock prompts for Oracle
     prompts = MockPrompts(
         {
-            r"Select the source": source_dict["netezza"],
-            r"Enter Secret Scope name": scope_name,
+            r"Select the source": SOURCE_DICT["netezza"],
+            r"Enter Secret Scope name": SCOPE_NAME,
         }
     )
 
     with patch(
-        "databricks.labs.remorph.helpers.db_workspace_utils.DatabricksSecretsClient._scope_exists",
+        "databricks.labs.remorph.helpers.recon_config_utils.ReconConfigPrompts._scope_exists",
         return_value=True,
     ):
         recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
@@ -87,3 +78,70 @@ def test_configure_secrets_invalid_source(mock_workspace_client):
 
         with pytest.raises(ValueError, match="Source netezza is not yet configured..."):
             recon_conf.prompt_and_save_connection_details()
+
+
+def test_store_connection_secrets_exception(mock_workspace_client):
+    prompts = MockPrompts(
+        {
+            r"Do you want to overwrite `source_key`?": "no",
+        }
+    )
+
+    mock_workspace_client.secrets.get_secret.side_effect = ResourceDoesNotExist("Not Found")
+    mock_workspace_client.secrets.put_secret.side_effect = Exception("Timed out")
+
+    recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
+
+    with pytest.raises(Exception, match="Timed out"):
+        recon_conf.store_connection_secrets("scope_name", ("source", {"key": "value"}))
+
+
+def test_configure_secrets_no_scope(mock_workspace_client):
+    prompts = MockPrompts(
+        {
+            r"Select the source": SOURCE_DICT["snowflake"],
+            r"Enter Secret Scope name": SCOPE_NAME,
+            r"Do you want to create a new one?": "no",
+        }
+    )
+
+    mock_workspace_client.secrets.list_scopes.side_effect = [[SecretScope(name="scope_name")]]
+
+    recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
+    recon_conf.prompt_source()
+
+    with pytest.raises(SystemExit, match="Scope is needed to store Secrets in Databricks Workspace"):
+        recon_conf.prompt_and_save_connection_details()
+
+
+def test_configure_secrets_create_scope_exception(mock_workspace_client):
+    prompts = MockPrompts(
+        {
+            r"Select the source": SOURCE_DICT["snowflake"],
+            r"Enter Secret Scope name": SCOPE_NAME,
+            r"Do you want to create a new one?": "yes",
+        }
+    )
+
+    mock_workspace_client.secrets.list_scopes.side_effect = [[SecretScope(name="scope_name")]]
+    mock_workspace_client.secrets.create_scope.side_effect = Exception("Network Error")
+
+    recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
+    recon_conf.prompt_source()
+
+    with pytest.raises(Exception, match="Network Error"):
+        recon_conf.prompt_and_save_connection_details()
+
+
+def test_store_connection_secrets_overwrite(mock_workspace_client):
+    prompts = MockPrompts(
+        {
+            r"Do you want to overwrite `source_key`?": "no",
+        }
+    )
+
+    with patch(
+        "databricks.labs.remorph.helpers.recon_config_utils.ReconConfigPrompts._secret_key_exists", return_value=True
+    ):
+        recon_conf = ReconConfigPrompts(mock_workspace_client, prompts)
+        recon_conf.store_connection_secrets("scope_name", ("source", {"key": "value"}))
